Add lint and compile watcher. Move from tslint to eslint.
This commit adds a "grunt dev" task which signficiantly improves
productivity by watching the src/ directory and recompiling and
relinting files on save.

It also moves the linting from TSLint to ESLint. TSLint is being
deprecated, and ESLint is the recommended path forward. It also
has nicer integration with prettier for code formatting. gts, the
Typescript styles we extend, is also in the process of migrating,
so this commit updates to using gts@next which includes the eslint
config.