standalone: add rudimentary collapsing, start with files collapsed, make buttons smaller
Unfortunately, still can't lazily create DOM subtrees, so page loading
is still very slow. It's not easy to do so, because the "run subtree"
button (as implemented) depends on the tree already being processed.
(However it could be done without generating the actual DOM.)
This page needs more serious overhauling.