Fixes about hasDynamicOffset and multisampled in BGL
hasDynamicOffset and multisampled should be undefined in BGL if the
binding types are incorrect. Previously we allow them to be false.

This change moves hasDynamicOffset tests out of visiblity test
and combine it with other optional parameters in BGL.

It also added a validation rule that if the binding type is storage
texture, the viewDimension can't be 'cube' or 'cube-array'.