Fix a bug for texture usage validation in compute
Resource usage validation scope in compute is defined by dispatch calls.
If dispatch command is not called in a compute pass, we don't need to
do resource usage validation for that pass. As a result, it should success
in this situation.

This change also sets compute pipeline and calls dispatch command for
compute pass if the test is not testing validation scope.