standalone: Pass parent element down instead of returning child element up
in the recursive DOM generation code.

This makes it so we can attach elements to the passed-in parent
element at any time, instead of having to generate DOM immediately.

Also default to expanding all the way down to the test level on load
(instead of only down to the file level), hiding only cases by default.