standalone: Don't generate DOM when test finishes (#323)
* standalone: Don't generate DOM when test finishes

Previously the test status and logs would get inserted into the DOM
immediately when run. This just stores off the result so that it can be
rendered into the DOM on-demand.

* standalone: Fix "expand all" button

Previously this just searched the entire page for checkboxes and checked
all of them. With lazy DOM generation, this didn't work anymore.

Returns a function from GenerateSubtreeHTML whichs sets the subtree as
"checked", recursively.

* hide testcaselogs by default

* nit