Break gen_listings dep on ts-node / fast-glob. Add src_dir/out_dir args (#492)
* Break gen_listings dependency ts-node / fast-glob. Add src_dir/out_dir args

Previous to this patch, it was necessary to run gen_listings from the
root of the source tree using ts-node. These requirements make it
necessary for listing generation to take extra dependencies which may
be problematic in environments where the set of dependencies are vendored
and downloading many dependencies with npm is undesirable. Even with no
dependencies, running from the root of the source tree means the
transpiled .js file can't be run directly.

This patch adds a tsconfig for transpiling tools/ for usage directly
with Node.js, and a build step in the grunt presubmit check for it.
Transpiling tools/ without extra dependencies on the actual test
module sources requires adding an extra argument to gen_listings to skip
test module validation, as well as removing the test description from the
listing (fixes #391).