Support passing skip/fail test expectations in wpt or by command line (#509)
* Support passing skip/fail test expectations in wpt or by command line

On the command line, test expectations are accepted by a command line
flag --expectations path/to/expectations.ts, where expectations.ts is
a module that exports RawTestQueryStringWithExpectation[] as
"expectations". In WPT, the page should define a global variable
|loadWebGPUExpectations| which is a
Promise<RawTestQueryStringWithExpectation[]>.

The test expectations are filtered down to only what is relevant for
the running test query, and applied to the testcases before they run.
A 'skip' expectation will entirely skip a testcase, and a 'fail'
expectation will surface failures as passes. If a 'fail' expectation
is found, but a test actually passes, the harness will log an warning
for the unexpected pass.

* Update tools/deno

Co-authored-by: Luca Casonato <lucacasonato@yahoo.com>

* Address comments from code review

* testcaseQuery -> filterQuery

Co-authored-by: Luca Casonato <lucacasonato@yahoo.com>