Remove it_is_invalid_to_use_a_texture_view_created_from_a_destroyed_texture
This test was wrong because it was expecting .finish() to have a validation
error instead of GPUQueue.submit(). Replace the test with a simple
texture_state test that checks it is valid to create a texture view
from a valid/destroyed texture, but invalid with an error texture.

Tests that check behavior of destroyed textures in finish() and submit() will
be moved/written along with tests that test specific commands that use
textures.

Also, replace inline lists of ['valid', 'invalid', 'destroyed'] with a
constant `kResourceStates`.