Introduce expectGPUBufferRepeatsSingleValue (#723)
* Introduce expectGPUBufferRepeatsSingleValue

This helper offloads work onto the GPU if the buffer is large enough,
rather than the typical behavior of reading back values to validate on
the CPU. This makes it more suitable for checking the contents of large
buffers quickly.

This is then used by expectSingleColor to validate its buffer of
copied texture data. For an empirical comparison, verifying the contents
of a 4k texture could take ~7-10 seconds before this change, and with
this change it takes only a few milliseconds.