Clamp limits between 0 and their maximum IDL value
requestDevice_limits.spec.ts was sometimes producing negative values
if the parameterization subtracted a value larger than the supported
limit value itself. Fix this by clamping to 0.
Also, the test clamped always to the maximum unsigned long value which
is incorrect for limits that are unsigned long long. Fix this to clamp
to Number.MAX_SAFE_INTEGER for limits that can be larger.