webgpu/shader/exection/builtin: Fix checkExpectation()
The `expected` field of a case is an array of possible allowed values. The logic was checking that the output value matched *all* the values in the array, which would never work.

Use the `ansi-colors` package to colorize the expected output. This significantly reduces the mental-parsing of a failure message.