webgpu/shader/exection/builtin: Fix checkExpectation() (#781)
The `expected` field of a case is an array of possible allowed values. The logic was checking that the output value matched *all* the values in the array, which would never work.

Fix the subnormal cases of the `abs()` tests. These are allowed to return 0.

Also simplify the formatted string for 0, Infinity, -Infinity.