Show some loading progress
It's a little disconcerting that the live CTS takes so long to load.
Here's one idea for putting something in that would let some progress
show.

Maybe we should also not add the buttons, or make them unclickable,
until loading has finished?

Note: I was going to use CustomEvent instead of MessageEvent but
CustomEvent does not exist in node 😭