Fix RenderBundle read-only state mismatch test
Fixes an erronious interpretation of the spec text. States don't
need to match exactly between the render bundle and the pass, but
the render bundle needs to be at least as strict as the pass.

Also, the test was creating invalid passes/bundles for combined
dpeth/stencil formats. Those are now filtered out.