zero_init: add constant zero to prevent driver unrolling loops.
The shader execution zero init tests produce large nested loops on
constant indices that are very tempting for drivers to unroll. This
commit prevents unrolling by adding zero to the loop bound using a zero
coming from a uniform buffer. This is not known to the shader compiler
which then can't optimize as much.