Add an optional before fn which is called before the test runs
In the following commit, the before function will be run
once for each testcase, but not for subcases inside that
testcase. This allows some common code to be performed
ahead of time and shared between subcases. This is
immediately useful so that device selection out of the
device pool can be performed once and then shared for
all of the subcases. Subcases should not use different
devices and will be refactored to select the device
in the `before` function, without relying on subparams.