Move test device selection to an optional before clause
The before function is an optional function which will be
run once for each testcase, but not for subcases inside that
testcase. This allows some common code to be performed
ahead of time and shared between subcases. This is
immediately useful so that device selection out of the
device pool can be performed once and then shared for
all of the subcases. Subcases should not use different
devices and will be refactored to select the device
in the `before` function, without relying on subparams.

The before function operates on a FixtureSharedState
which each Fixture class may override. This state object
is shared between all fixtures created for subcases within
a single testcase.

The test device is stored in this shared state so that now,
reserving a device from the device pool occurs once for all
subcases, it is shared between the subcases, and released at
the end of the testcase.

This means that if there are unexpected validation, OOM, or
device lost errors, they will only be reported at testcase
granularity, not subcase granularity.