Ensure buffer won't be reset in subresources,buffer_usage_in_render_pass
This patch improves the test subresources,buffer_usage_in_render_pass
to ensure all the commands won't reset a buffer in the render pass
encoder. The situations that a buffer is reset by another call in one
render pass will be tested in another test.

This patch also shortens several test parameters to ensure the name of the
test won't be too long to handle on the bots.