Explicitly destroy devices created by requestDevice_limits
These tests are consistently failing after N iterations with
device creation failures (for example, with D3D12 it starts
reporting that createCommandQueue() failed with E_OUTOFMEMORY).
Local testing shows that if a garbage collect is forced the tests
start passing again, indicating that the issue is that the devices
are not being cleaned up fast enough.

Explicitly destroying the new device after each test iteration
appears to resolve the problem.