Split WGSL test files
This PR creates a shader/validation/parse folder and moves the WGSL
parsing tests into the folder. The tests are split out by type to
make the test files more targeted.

For the var/let cases half of them are in the shader_io folder where
they relate to IO sharable types.