Extend WGSL interpolate tests.
The current interpolation tests will not attempt to validate an
interpolation with just the type (e.g. `@interpolate(flat)`). This is
due to the `,` always being appended so `@interpolate(flat, )` is
generated which is not a valid value.

This PR updates the generation to only add the `,` if there is a
sampling value to be appended. As well, it makes the space variable so
we can validate that the space between the `,` and the sampling isn't
required.

Issue: #1135