Remove uniqueId and plan_autogen from WGSL tests. (#1219)
We're currently evaluating if the bikeshed assert functionality will provide
us with the required unique identifiers. Until that happens, these old identifers
will not be used.

This PR removes the `uniqueId` field, it can be re-added if we determine assert
is what we need. The `plan_autogen.md` file is also removed as we are not currently
autogenerating the test files for WGSL. This can also be re-added if we determine the
bikeshed assert functionality will work for us.