Remove old builtin combined function file.
This PR removes the old single combined builtin function test file which contains
unimplemented tests for several builtins. The list in this file is out of date with
the spec, the spec descriptions are out of date as well.

Each of these missing specs will get filed as a CTS bug and tracked through the CTS.
When added back they should each get an individual test file.