wgsl: Add support for f64 to Scalar
This is needed so that test expectation can be specified in a higher
precision than what WGSL is operating on (64-bits for 32-bits). This
allows for fluent specification of test conditions like correctly
rounded, instead of having to implement a per case Comparator.

This PR also renames some of the related helpers with similar names so
it is clearer what they return.

Fixes #1332