Remove dupes in anyOf to clean up error messages (#1401)
The function returned by anyOf no longer adds duplicate expected values
to the failed list.

Also, many tests call calculateFlushedResults() that was returning a
Set<Scalar> of usually the same two Scalars (most values to not flush to
0). However, the set would always contain both values because Scalar is
a reference type, so two Scalar instances of the same value do not
compare equal (see https://stackoverflow.com/a/29759699 for why we can't
use Sets this way right now). Instead, calculateFlushedResults now
returns an Array<Scalar> to make it obvious that Scalars of the same
value are not being dropped (and Array is likely more efficient than
Set).