Duplicate params: catch in ParamsBuilder, improve error in loadTreeForQuery
The check in ParamsBuilder is a bit more local than the one in
loadTreeForQuery, so hopefully easier to debug.

Only combine()/combineWithParams() can catch duplicate params
efficiently (just once, instead of once per parent parameterization), so
other methods don't check for them.
In those cases, the assertion in loadTreeForQuery will be hit, so
improve the error for that assertion, too.