Implement handling non-monotonic functions
Adds in machinary for specifying where extrema occur for non-monotonic
functions. This is achieved via seperating the API that is called by
tests and the implementation to allow for an additional optional
extrema function to be defined.

This allows moving the atan2 complexity related to multiple return
values @ y = 0 from the test into the framework implementation, which
will help when using atan2 to declare other tests.

I also cleaned up a bunch of little naming things that my IDE was
bugging me about.