Implement new F32Interval class for floating point tests
This implements the F32Interval class and fundamental accuracy
functions described in my prototype for reworking,
https://github.com/gpuweb/cts/pull/1402, how floating point tests are
written.

This also adds in testing that did not appear in my prototype.

There are a couple of functions that are not currently used that I
have included in this patch that have suppressions for unused-vars on
them. I included them in this patch since they are logically grouped
with this code. These supressions will be removed in future patches.

Issue #792