Change Scalar.toString() to include hex bits
Currently this is include the value of the Scalar in base 16, which
isn't that useful. For unsigned integers this maps directly to the
bits, but for pretty much any other numeric type it is just making a
harder to read version of toString().

I am pretty sure the intent here was to include the bits in a readable
format so that one could easily evaluate what is being stored.

This changes the return of toString from:

-1.00048828125(0x-1.002, subnormal: false)

to this:

-1.00048828125 (0xbf801000)

'subnormal' is included after the hex value if the value is subnormal