Implement new F32Interval class for floating point tests (#1500)
This implements the F32Interval class and fundamental accuracy
functions described in my prototype,
https://github.com/gpuweb/cts/pull/1402, for reworking how floating
point tests are written.

This also adds in testing that did not appear in my prototype.

There are a couple of functions that are not currently used that I
have included in this patch that have suppressions for unused-vars on
them. I included them in this patch since they are logically grouped
with this code. These surpressions will be removed in future patches.

Review discovered a bug in oneULP's implementation, which is fixed in
this PR. 

Issue #792