Seperate out interval operation definitions from builder functions
For all of the F32Interval builder functions, factor out the
associated operation structure definition. This prevents the subtle
error where a variable being referenced in an arrow function doesn't
actually exist in that scope, but does exist in the surrounding scope,
so is silently bound in.

There are a number of cases, like logInterval, log2Interval, and
ulpInterval where with bug existed and has been corrected. Tests have
been updated and simplified to be correct and more human readable.

Fixes #1625