Add support for multiple intervals to FP testing framework
This change handles situations that require multiple discrete
intervals by having the test specify multiple interval generators when
invoking make*Case. This means that for cases like clamp, there will
need to be two generators, clampMedianInterval and clampMinMaxInterval
that will be passed in.

The upshot is that the code and interfaces in f32_interval.ts do not
need to handle multiple interval returns, so the complexity is
isolated to expression.ts, which should be simplified once the old
framework code can be removed.

Fixes #1641