Update f32 tests to handle returning undefined for infinity
The WGSL spec has been clarified to allow implementations to return
undefined values when they do not implement handling infinity. For CTS
tests this means that anytime infinity is encountered it may
potentially turn into an undefined value. In general terms this means
that acceptance intervals for certains cases need to loosened, and a
bunch of corner cases related to clamping infinities can be removed.

This is a big monolithic change, because the alternate would be
creating a parallel framework infrastructure with the new behaviour,
and rewritting tests piecemeal to use the new framework. Which I am
currently already in the process of doing wrt migration to the
interval framework. Having 3 different testing frameworks live at the
same time seems like a recipe for disaster.

Highlights of this change:
- F32Interval.infinity() has been renamed F32Interval.undefined()
- Infinity clamping behaviour in the F32Interval constructor is gone
- *ToIntervalOp.impl is no longer responsible for checking if inputs
  are valid
- *ToIntervalOp.domain? property has been added to allow for operation
  definitions to specify addition contraints on the inputs, run*Op
  actually tests these limits
- run*Op is now responsible for checking that inputs/outputs of
  operations are finite, otherwise returning undefined
- Usages of Number.POSITIVE/NEGATIVE_INFINITY on the input side of
  unittests related to generating intervals have been change to
  kValue.f32.infinity.positive/negative, for consistenancy and since
  Number.* is a f64 value

Fixes #1642