Update f32 tests to handle returning undefined for infinity (#1655)
The WGSL spec has been clarified to allow implementations to return
undefined values when they do not implement handling infinity. For CTS
tests this means that anytime infinity is encountered it may
potentially turn into an undefined value. In general terms this means
that acceptance intervals for certains cases need to loosened, and a
bunch of corner cases related to clamping infinities can be removed.

This is a big monolithic change, because the alternate would be
creating a parallel framework infrastructure with the new behaviour,
and rewriting tests piecemeal to use the new framework. Which I am
currently already in the process of doing wrt migration to the
interval framework. Having 3 different testing frameworks live at the
same time seems like a recipe for disaster.

Fixes #1642