Make F32Interval unittests more idiomatic
Adds a type IntervalBounds to represent the [number]|[number,number]
that can be passed into the F32Interval constructor. This is
propegated to f32_interval.spec.ts to replace how bounds are
represented there.

This allows for replacing duplicated bounds, i.e. [0, 0] with [0].

Additionally there is a bunch of cleanup done with having common
error handling and removing patterns where elements of t.params are
being unpack into a variable, but never transformed.

Fixes #1762
Issue #1747