Add storage_texture,formats test to createBindGroupLayout.spec.ts
According to the specification, the storageTextureLayout.format must
be a format which can support storage usage if the storageTextureLayout
is not undefined. So this PR adds a new test to ensure that a validation
error is generated if a format used by the storageTextureLayout doesn't
support the storage usage.

Issue: #885