webgpu/shader/execution: Check validation before testing results
Check that the shaders compiled, pipelines created, workgroups dispatched, etc, before attempting to inspect the shader output.

For shaders that are not currently compiling, this massively reduces the amount of error spam, as we no longer spew an error per failed expression case, or which some calls to `run()` contain thousands.