Linearly spread `f32FullRange` values over possible f32 values (#1786)
This changes the generation code to linearly spread over the possible
values, i.e. take every nth value, instead of spread based on the
range. This prevents all of the values from being bunched up near the
ends, since the range of f32 is orders of magnitude larger than the
number for valid f32 values.

Also fixed the order of the negative normals.

Fixes #1782