Fix a wrong bit operation to decide the validation in indirect_dispatch_buffer,usage (#1819)
The indirect_dispatch_buffer,usage test has been deciding the validation of
the buffer usage through OR operator between INDIRECT and |bufferUsage|.
The operation has decided that all usages are valid and it's not the purpose
of the test. So this PR replaces `|` operator with `&` operator to fix
the problem.