shader/execution: Fix shader creation errors
With https://github.com/gpuweb/gpuweb/pull/3400, constant expression indices that are OOBs are a shader-creation time error.

Fix the robustness tests by placing the indices into a `let`, which forces the index to be a runtime-expression, avoiding the error.

Also fix the WGSL indentation, making the shader code easier to read.