Add 'buffer,resource_binding_size' to createBindGroup.spec.ts
The specification says that the resource binding size should be equal to or less than
limits.'maxUniformBufferBindingSize|maxStorageBufferBindingSize' if the BindGroup
entry defines buffer and the buffer type is 'uniform|storage|read-only-storage'.

So this PR adds a test to check if a validation error is generated if the binding
size is greater than limits.'maxUniformBufferBindingSize|maxStorageBufferBindingSize'.

Issue: #884