Add a 'storage_texture,format' test to createBindGroup.spec.ts
The specification says that the format of the storage texture
should be equal to the resource's descriptor format if the BindGroup
entry defines storageTexture.

So, this PR adds a new test to check if a validation error is generated
if the storage texture's format is not equal to the resource's descriptor
format.

Issue: #884