Add execution tests for logical binary operations
This covers `and` (`&`), `and` with short-circuiting (`&&`),
`or`(`|`), and `or` with short-circuiting (`||`)

Short circuiting is tested in existing tests, which are referenced in
the source file.

This also required changes to how the expression builder operated to
bind the `!= 0u` for converting bools coming out of storage to just
the param.

Issue #1868