Change from using `full` to `dense` for utility function names
The utility function in question generate values accross the entire
range of a numeric type, but do not generate every value in the range,
so it can be ambigious when reading the name if it is going to produce
all possible values or a sample of values.

`dense` is clearer that it is a larger selection of values than the
corresponding `sparse` functions.