Convert asin and acos to a filtered range when needed.
When doing const-eval for the `asin` and `acos` builtins they are
only defined in the `[-1, 1]` range. Tests outside that range fail
with a validation error. When not using const-evaluation having
values outside the range is useful for validation.

This CL adds a `sourceFilteredF32Range` which will clamp the F32
values if the source is `const`. The `asin` and `acos` tests are
updated to use the new filter.