op: Implement 'index_format,change_pipeline_after_setIndexBuffer' test in index_format.spec.ts
This PR adds a new test to ensure that setting the index buffer before setting
the pipeline works well. This is important for backedns where the index format
is passed inside the call to setIndexBuffer because it needs to be done lazily
to query the foramt from last pipeline.

Issue: #1968