Simplify code around ULP calculations and fix subtly wrong tests
Due to multiple layers of control FTZ behaviour was not controlled at
the top-level during some unittests, so FTZ was always assumed, which
lead to tests with subtle ordering errors to be passing.

This patch simplifies the calling stack for ULP calculations, and
allows tests to plumb through control of flushing behaviour as
needed. The incorrect test is corrected.

Fixes #1971