WIP: Remove infinities from values for f32 tests
This patch removes the use of infinities in inputs for all of the f32
tests. It is also introduces a framework for specifying filtering of
inputs to const-eval tests, so that inputs that would cause validation
errors can be avoided.

This patch currently does not have tests covering the new
*LargestIntermediateValue functions, those will be added before final
review and merge.

This does NOT remove the generation of NaN/Infinity via 1/0 when
writting out testing values. That will be handled in a follow on PR,
since that may have impact outside of the numeric tests.

Issue #2020