Replace usage of vector inputs with fullF32Range for arithmetic tests
During one of the refactorings of this code, the vectorTestValues were
substituted in for fullF32Range when generating the inputs. This is
not desired, so I am reverting this behavioural change.

There are more broken const-eval tests on some backends due to this
change, though const-eval was failing on these backends already. This
will be resolved in a pending change I have for filtering out bad
const-eval inputs, https://github.com/gpuweb/cts/pull/2045.

I am sending up this patch early, so that coverage of division can be
restored for investigating potential issues with the atan2 tests.

Issue #2020