Remove infinities from all f32 test inputs
This will reduce testing failures due infinities being past into
const-eval tests, which cannot be represented.

Some const-eval tests still fail due to intermediate calculation
values not being representable in f32. Filtering inputs to prevent
this will be done in a future patch.

This patch does not remove the `1/0` trick used in Scalars to attempt
to output infinity. That will be removed in a future patch.

Issue #2020