Implement generate*Cases functions and use them
This replaces the various patterns used through out the numeric tests
to create an array of input tuples and then maps them to make*Case to
a common pattern that uses cartesianProduct for calculating the
tuples.

This reduces duplication of code and will allow for easier future work
on filtering cases for const/non-const by having a single point where
the filtering needs to occur.

Issue #2058