Add flag to unroll const-eval loops
FXC is extremely slow to compile a number of individual assignments to a storage buffer.

MSL is extremely slow to compile a for loop reading from a constant array.

Add a flag to customize the behavior, so we can optimize CTS runs for the given platform.

Note: Exercising different code paths based on platform is far from ideal, however:
• This is on a flag, so the runner can opt into this or not.
• The unrolling only applies to const-eval tests, which primarily exist to check that constant evaluation produces the correct answer. Whether assignments are done via a loop or not should not change this behaviour.