Unclutter the URL
The URL is getting long as new options are added. It's common to want
to edit the q parameter but as more options are added this gets harder
as it get pushed off the to the right.

So, only show options that are non-falsy but, display the options
on the page and make them editable.