Test texture descriptor validation in GPUCanvasContext
The specification describes that the configure() and getCurrentTexture()
both run GPUTextureDescriptor validation steps on the descriptor for the
canvas textures. The only things in control of the application are:

 - the format (already tested)
 - the view formats
 - the size (through canvas.width/height)

This commit adds tests for the view formats and the size, checking for
corner cases where the size is 0 before the configure() or becomes 0
after the configure.