Add a ResizeObserve test
This is a test I submitted to the Web Platform Tests but changed
to use WebGPU instead of canvas 2d.

It works by creating a bunch of divs(ref) or canvas(test)
and setting them to be various percent sizes.

For the divs it sets their background-image pattern so that 1 pixel
in the pattern is one device pixel. This is easy since it's just
patterSize / devicePixelRatio

For the canvas it uses ResizeObserver and given the size it receives
it fulls each canvas with a 1 canvs pixel to 1 device pixel pattern.

The 2 pages should match. If ResizeObserver is bad they won't.

The pattern used is 4x4 which gives more chances for things to be
off of alignment than a 2x2.