Maximize the acceptance interval size for `dot` (#2151)
`dot` currently calculates the interval size by summing the
multipltications in element order, but the spec does not require a
specific operation ordering. The error in the calculation is dependent
on the specific ordering.

The new implementation sums all of the negative and positive terms
together seperately from each other, since the largest intermediate
value, thus error, will occur from one of these two intermediate
results. This avoids needing to calculate all of the permutations of
the addition ordering to discover the largest acceptance interval.

Issue #2134