Remove phantom dependency on reftest-wait.js, copy code into CTS
We were relying on takeScreenshotDelayed to be defined by WPT, but it's
not present when running test pages outside WPT, which means there are
always errors, which makes it hard to see when we've forgotten the
script tag to import it.

The implementation is very simple (just removes a class from <html>), so
just copy it in.