Emit `f32` values for F32Interval.toString()
The existing version of toString will emit the TS `number` as a
string, which will just be the decimal numeric value.

In test logs it is normally more useful to have the value converted to
`f32`, since that includes the hex encoding.

This format is what is already used for the input and returned values
in the logs. This should only be changing expectation strings.

Fixes #2295