shader/execution: Port evaluation order tests to flow_control
The `flow_control` framework helps with test readability, the errors are easier to understand, and we can add test case permutations, like to allow / disallow constant folding.

This is not an exact porting - some tests have fewer permutations, but I've tried to cover a wider range of interesting cases, like compound assignments and increments.