fix(gen_wpt_cts_html): use non-zero exit code on incorrect arg count
The `gen_wpt_cts_html` script uses some simple checks against `argv`'s
length to ensure that input is valid. When that arg count does not match
expected usage, the tool prints instructions with examples of expected
usage, and returns an exit code of 0. This is a problem, though, because
0 conventionally indicates that the program ran without issue; by
definition, we received incorrect input! This is definitely a case where
it should return a non-zero exit code to indicate failure.

Fix this by returning 1 instead of 0. Easy peasy!