Refactor how floating point intervals are calculated (#2468)
Introduces a FPTraits class that things like limits, constants,
utilities, and interval generators exist in. This reorganizes the
existing collection of free functions that all loosely depend on
F32Interval into a tighter structure where they pull information about
how to build intervals and what the limits for floating point numbers
are from their context.

There is currently one concrete implementation of FPTraits class for
f32, there will be more in the future to support abstrate float and
f16.

There is currently code in the generic FPTraits class that is f32
specific, like magic numbers for precision. As things like abstract
float and f16 are implemented I expect there to be refactorings to
move these things around, either adding more constants, or moving
specific interval generators down to specific floating point
implementations.

All of the implementation code has been moved to floating_point.ts

f32_interval.ts contains pass-through shims that allow existing code
and tests to use the old API for generating intervals. Future patches
will work through migrating the rest of the CTS off of the old
API. Currently testing for the interval generators runs through the
shim, so that the codepath being used in CTS is under test, and it
reduces the diff for this patch.

Most of the existing code outside of the interval generators doesn't
need to change at this point, expect for specific locations like
serialization/deserialization, where FPInterval (formally F32Interval)
is used explicitly. This is due to FPInterval being an actual class vs
type/interface, so it it is non-trivial to alias it in the shim. So
for the few places where the refactoring needed to leak out it has,
instead of going through heroics for conceptual purity.

Issue #2416