Migrate unpack* tests to new framework
These are pretty tightly coupled, so migrated them all in one patch.

The are in F32Traits instead of FPTraits, since they are only defined
for f32.

There are a couple of other tests that are f32 only I think
that need to be moved down to the leaf class like this, but I will do
that when I start implementing Abstract Float.

Issue: #2416