Implement core abstract float support
This PR adds in the various utilities needed for abstract float (f64),
i.e. nextAfter, ULP, serialize/deserialize interval.

An implementation FPTraits, FPAbstractTraits is implemented to use
these utiliites.

Test coverage is added for all of this new code.

The interval generators are roughed in enough to instantiate the
FPTraits object, but not tested or used. Further PRs will implement
testing for this code and add it being used.

Fixes #2416