Refactor FPInterval and utilities unittests with traits (#2600)
This PR refactor the FPInterval and Utilities unittests for floating point, merging the splited tests for f32 and abstract float into subcases of different trait. In this way we can reuse the (mostly) duplicated cases for f32 and abstract float, and make it easier to add the f16 cases for these test (just adding another f16 trait will work).