Make need for serialization of Comparators more obvious (#2601)
* Make need for serialization of Comparators more obvious

As-is the CTS framework just requires you to define a function to
create a Comparator and pass it as a Expectation in a Case. This
causes issues if the writer then attempts to implement using the
CaseCache, since that requires the Comparator to have defined
serialization and deserialization. There is no type level enforcement
of this, so errors are not detected until runtime with the cache
enabled.

This PR reworks the Comparator type to have an impl function and kind
tag as required properties, along with optional data for
serialization. This forces the user to actively bypass the type system
to define a Comparator without serialization. Additional documentation
has been added around the relevant definitions to point the user in
the correct direction.

This PR doesn't change the entire type system to strictly require
serialization/deserialization. This is because some of the internals
of the testing framework use Comparators as part of building and
running tests.

One could add in serialization/deserialization code to these
Comparators, but this would allow silent errors when internal
components leak out into the cache unintentionally. The other option
is to split the current Case/Expectation system to have Serializable
and non-Serializable implementations. This creates a bunch of
duplicated code and significant amounts of type conversion for a bit
of intellectual cleanliness.

This implementation chooses to make the framework user's error a type
level error, so detectable by tooling, since there is likely to be
many more changes to code defining tests and using the framework vs
the framework itself. Issues in the framework will still cause runtime
issues, and thus integration failures, but hopefully localize them to
code that is changing less frequently.

The bitcast tests are given an explicit opt-out from the
serialization. A follow up PR will rewrite them to properly define
their Comparators and remove this legacy opt-out.

Issue #2593